[WIP] split api into routes.go issue #2391
